Split SME chat list and add auto-scroll handling
- Move message rendering into a memoized list component
- Add scroll-aware auto-follow behavior during streaming and send states
- Render streaming text as plain text and simplify bubble updates
